refactored file name add folder tasks
