+++ README.md
@@ -1,13 +1,13 @@
 # python_tasks
 
 
-| Name | task 1 | task 2 | task 3 | task 4 | task 5 | unuttest |
+| Name | task 1 | task 2 | task 3 | task 4 | task 5 | unittest |
 |----------|:-------------:|:-------------:|:-------------:|:-------------:|:-------------:|------|
-| Oleksandr Makar | 86.а | 107 | 224 | 323 | 554 | Kseniia Malinovska |
-| Ostap Halapats | 86.б | 108 | 225 | 325 | 555 | Taras Stetsiak |
-| Yurii Vantukh | 87 | 178.б | 226 | 329 | 559 | Olga Hasyuk |
+| Oleksandr Makar | 86.а | 107 | 224 | 323 | 554 | Nazar Shchur |
+| Ostap Halapats | 86.б | 108 | 225 | 325 | 555 | Serhii Sibirtsev |
+| Yurii Vantukh | 87 | 178.б | 226 | 329 | 559 | Kseniia Malinovska |
 | Nazar Shchur | 88.а | 178.в | 227 | 330 | 560 | Oleksandr Makar |
-| Serhii Sibirtsev | 88.б | 178.г | 243.а | 331.а | 561 | Ostap Halapats |
-| Kseniia Malinovska | 88.в | 178.д | 243.б | 331.б | 562 | Yurii Vantukh |
-| Taras Stetsiak | 88.г | 182 | 241 | 332 | 568 | Nazar Shchur |
-| Olga Hasyuk | 86.г | 184 | 242 | 340 | 569 | Serhii Sibirtsev |
+| Serhii Sibirtsev | 88.б | 178.г | 243.а | 331.а | 561 | Olga Hasyuk |
+| Kseniia Malinovska | 88.в | 178.д | 243.б | 331.б | 562 | Taras Stetsiak |
+| Taras Stetsiak | 88.г | 182 | 241 | 332 | 568 | Ostap Halapats |
+| Olga Hasyuk | 86.г | 184 | 242 | 340 | 569 | Yurii Vantukh |
