+++ tasks/task_332.py
@@ -1,6 +1,6 @@
 """"Equate lagrange's theorem for given number"""
 
-from tasks.task_88r import input_and_validate
+from tasks.task_88d import input_and_validate
 
 
 def lagrange(natural):
